+++ .vscode/launch.json
@@ -11,7 +11,7 @@
       "program": "coach.py",
       "console": "integratedTerminal",
       "envFile": "${workspaceFolder}/.env",
-      "args": ["--verbose"],
+
       "env": {
         "PYDEVD_DISABLE_FILE_VALIDATION": "1"
       }
@@ -23,7 +23,7 @@
       "program": "coach.py",
       "console": "integratedTerminal",
       "envFile": "${workspaceFolder}/.env",
-      "args": ["--verbose"],
+
       "env": {
         "PYDEVD_DISABLE_FILE_VALIDATION": "1",
         "AICOACH_audiomode": "text",
@@ -37,7 +37,7 @@
       "program": "coach.py",
       "console": "integratedTerminal",
       "envFile": "${workspaceFolder}/.env",
-      "args": ["--verbose"],
+
       "env": {
         "PYDEVD_DISABLE_FILE_VALIDATION": "1",
         "AICOACH_aibackend": "Mocked"
@@ -60,6 +60,15 @@
       "console": "integratedTerminal",
       "envFile": "${workspaceFolder}/.env",
       "args": ["--debug", "--clean", "sync"]
+    },
+    {
+      "name": "CLI add players",
+      "type": "debugpy",
+      "request": "launch",
+      "program": "repcli.py",
+      "console": "integratedTerminal",
+      "envFile": "${workspaceFolder}/.env",
+      "args": ["--debug", "addplayers", "--from=2024-06-07"]
     },
     {
       "name": "CLI deamon",
