--- replays/__init__.py
@@ -1,5 +1,0 @@
-from .reader import ReplayReader
-from .util import time2secs
-from .db import engine as db
-from .db import replaydb
-from .newreplay import NewReplayScanner
